Complete ASM

+++ BusinessObject/Products.cs
@@ -14,9 +14,9 @@ namespace BusinessObject
         public int CategoryID { get; set; }
         public string QuantityPerUnit { get; set; }
         public decimal UnitPrice { get; set; }
-        public short UnitsInStock { get; set; }
-        public short UnitsOnOrder { get; set; }
-        public short ReorderLevel { get; set; }
+        public int UnitsInStock { get; set; }
+        public int UnitsOnOrder { get; set; }
+        public int ReorderLevel { get; set; }
         public bool Discontinued { get; set; }
 
     }
